Suggestion: Combine package display name & release tag base name
We were tracking 2 almost-identical pieces of info throughout the scripts and configs for basically no real benefit, other than optionally getting a slightly more human-friendly name in the dropdown when you build a preconfigured package from that pipeline.  I merged these into one "PackageName" variable that we pass around.  This serves the purpose as both the display name that you pass into the "build-and-stage-preconfigured.ps1" script AND the base tag name that is used when publishing a release.