Add debug version
I am adding a debug version of this because snagit was having some issues with the add to dictionary feature in debug mode only.  I also found that this requires some additional changes to the scripts in this repo, as there are some assumptions made about where files are that are not true in debug mode (see next commits).