Include non-library files in the output lib dir for Mac
Previously, we were only including *.dylib and *.a in the output for this dir.  This adds an extra step to copy this files over as well.

The glib weirdly puts a header in this dir (lib/glib-2.0/include/glibconfig.h), which we need to be there on Mac.  It's possible other libraries are putting non-library stuff that we need here, so we'll just copy all non-lib files over from here all the time.