Fix infinite loop in Get-SymlinkChains if no symlinks found
There is typically a chain of symlinks to the actual file when dealing with these mac builds (ex. foo.dylib ==> foo-1.dylib ==> foo-1.2.dylib ==> foo-1.2.3.dylib).  The physical file is foo-1.2.3.dylib, but we want to just have the physical file be called foo.dylib and get rid of the symlinks (existing functionality in these scripts).  To get the right name for the file, we find all the physical files and then walk the symlink chain backwards until we get to the top.

There was a bug where if no symlinks are found from the Mac build, however, it was causing an infinite loop looking for them.  I changed this so that we will not try to walk the symlink chain to find the top-level symlink for each physical file, and just return the same name for the physical file and the top of chain file.