Consolidate common post-build steps for Mac
We were removing dylib symlinks in all of the post-build scripts for Mac, and we will now want to do this debug symbol stuff for all of the Mac builds, so just consolidate all of that in the Create-FinalizedMacArtifacts step, which runs before the post-build scripts.

This allows us to shorten up the Post-Build scripts and also remove a few of them.