Fix Remove-DylibSymlinks
We previously relied on all the symlinks being in a chain to the ultimate physical filename.

These would look like this:
foo.dylib (symlink) -> foo.1.23.dylib (symlink) -> foo.1.23.4.dylib
(actual file)

This worked for all of the libs we have had up to this point, but some libs no longer have all of these all in a chain like this.  Instead, some libs, such as the upcoming FFMpeg 6.1.1 we will be updating to soon, looks like this:
foo.dylib (symlink) -> foo.1.23.4.dylib (actual file)
foo.1.23.dylib (symlink) -> foo.1.23.4.dylib (actual file)

Instead of building up a chain, then, we do this:
1. Find all the physical files
2. Look through all the symlinks, and figure out which ones ultimately
   point to that physical file.
3. Choose the filename among the symlinks and actual files which has the
   shortest name (ex. foo.dylib)
4. Delete all the symlinks
5. Rename all of the physicial files to thier "short" versions, found in
   step 3
6. Update all of the metadata for what the dylibs are called and what
   filenames they link to, to use the short names

We were already doing a lot of this before.  The main change now is how
we choose which filename to keep, and how we determine which symlinks
are related to which actual files.