Fix missing distutils dependency issue
Python 3.12 no longer shipped with distutils: https://docs.python.org/3/whatsnew/3.12.html

Distutils was required on Mac for the vcpkg build of glib, however, which is required for the vcpkg build of pango.  We would get the error below when trying to build it:

ModuleNotFoundError: No module named 'distutils'

We were previously working around this by downgrading python to 3.11, where distutils is deprecated but still shipped with python.  This worked for awhile on our build agents, and then stopped working at some point, for reasons yet unknown.

There is another workaround for this issue, however, which is to install setuputils: https://stackoverflow.com/questions/77233855/why-did-i-get-an-error-modulenotfounderror-no-module-named-distutils

This uses this second workaround instead, which DOES work on our Mac build agents.