Refactor: Rename function "Check-RequiresEmscripten"
Renaming this to "Check-IsEmscriptenBuild", to more closely match the OS-checking functions.  While "Check-RequiresEmscripten" makes sense if we are only using it to determine if emscripten needs to be installed before running the build process, it sounds a little clunky from the perspective of a post-build script or anything after the time that we install emscripten.