Feedback: Update feature flags to control related encoders and decoders
Some feature flags enable the ability to use certain encoders and decoders.  To make it easier to control some of this in one place (preconfigured-pacakges.json), I have changed the logic so that the encoders and decoders we want to be enabled when the aom,dav1d,mp3lame,vorbis, and vpx feature flags are enabled are able to get enabled, now.

Normally, we would not need this when using vcpkg to do a build of FFMpeg, but since we are using --disable-everything first in portfile.cmake to make sure certain codecs don't sneak in unexpectedly, we need to update the cmake script to enable these codecs one-by-one.