Group feature flag sets to clean up preconfigured-packages.json
It is hard to see what the differences are between the sets of features we pass to the different ffmpeg builds, now.  This groups these sets of features a bit more logically in vcpkg.json, so that we can more easily see what the actual differences are between builds that are defined in preconfigured-packages.json.