Add script to generate version-info.json
This will generate the version-info.json file for use with Update-VersionInfoForDlls for Windows package post-build steps.

By default, many of the fields will probably be blank.  Vcpkg ports often do not add the version and copyright metadata to dlls, so we need to add that in after the fact, as necessary.