Updated versions
Some clients have installer issues if the file versions do not change when the dlls change.  To avoid this, I am manually appending ".100" as the 4th part of the version number, similar to what we do for FFmpeg when things like libmp3lame do not get updated.

This change has been made for:
- bin/brotlicommon.dll
- bin/brotlidec.dll
- bin/brotlienc.dll
- bin/bz2.dll
- bin/charset-1.dll
- bin/pthreadVC3.dll
- bin/pthreadVCE3.dll
- bin/pthreadVSE3.dll
- bin/zlib1.dll

Additionally, now that we have a script to generate this version-info.json file, I've added back the other files that are actually versioned correctly coming out of vcpkg, so that we can easily detect if those file versions have changed between builds of pango & friends.

I have done this for these libraries:
- bin/freetype.dll
- bin/gio-2.0-0.dll
- bin/glib-2.0-0.dll
- bin/gmodule-2.0-0.dll
- bin/gobject-2.0-0.dll
- bin/gthread-2.0-0.dll
- bin/iconv-2.dll <== also required a version update to .100!
- bin/intl-8.dll
- bin/libexpat.dll <== previously set version metadata successfully, but not other metadata
- bin/pango-1.0-0.dll
- bin/pangocairo-1.0-0.dll
- bin/pangoft2-1.0-0.dll
- bin/pangowin32-1.0-0.dll
- bin/pthreadVC3.dll <== also required a version update to .100!
- bin/pthreadVCE3.dll <== also required a version update to .100!
- bin/pthreadVSE3.dll <== also required a version update to .100!
- bin/zlib1.dll <== also required a version update to .100!