Update version: zlib1.dll
This builds with a version of it's own, but I had to override it for this build.

I had to increment the build number by 1, however, as this ends up being the same version as before.  If don't update this to a newer version, it won't work with our installers.