Make separation between "package" and "port" more clear
A "Package" is something that is defined in "preconfigured-packages.json".  A "Port" is the actual name of the port that the Package references.

A "Package" includes information about the port to run, the feature flags to run it with, the triplet, and other info of what to do when building it and staging it.

A "Port" and it's "Features" are a string like foo[feature1,feature2], which is what you would call if you were simply using vcpkg directly (ex. `vcpkg install foo[feature1,feature2]`).

I also moved the invoke-build.ps1 to scripts, to make it more obvious which script we prefer users to use from the commandline.