Commit d188c9d2364efb76ad42fe0122a5725d7878ebbf has in the comments that it pulled updates from the 2025.06.13 tag of vcpkg, but that tag looks like it has libzip 1.11.3 in it: https://github.com/microsoft/vcpkg/blob/2025.06.13/ports/libzip/vcpkg.json#L3, not 1.11.4.
It appears that libzip 1.11.4 first appeared in the 2025.07.25 tag: https://github.com/microsoft/vcpkg/blob/dd3097e305afa53f7b4312371f62058d2e665320/ports/libzip/vcpkg.json#L3

It also is in the more recent 2025.08.27 tag: https://github.com/microsoft/vcpkg/blob/120deac3062162151622ca4860575a33844ba10b/ports/libzip/vcpkg.json#L3

I diffed the files in both the tags for this port, and both had the identical 1.11.4 port.  I'm updating this to make the copy of vcpkg be off of the 2025.08.27 port, just to make sure we are pointing to the latest stable bits for the vcpkg tooling and any ports we may want to use as features of libzip, so that they will use the latest version(s) of those dependency port(s).

Not that "vcpkgHash" can either be the commit hash in the vcpkg repo you want to use (ex. 120deac3062162151622ca4860575a33844ba10b) or it can be the tag name you want to use, which is a little cleaner-looking, and ensures you are likely on a "stable" version of vcpkg (ex. 2025.08.27).