Add aom to custom-ports
We have to add this due to a bug in build/cmake/aom_optimization.cmake.  It is not detecting that nasm has multi-pass optimization support on Mac.  This bug has since been fixed in their "main" branch, but they have not cut a release tag with the fix yet, and the vcpkg port is pointing to their latest release tag (that does not include this fix), which is 3.13.1.