Add aom patch to fix nasm multi-pass optimization capability check
This should fix this error on Mac:
CMake Error at build/cmake/aom_optimization.cmake:219 (message):
  Unsupported nasm: multipass optimization not supported.

This patch previously worked with macos-14 image version 20251016.436, but this patch seems to be needed for macos-14 image version 20251104.0148.

I also found that there was a similar patch on the main branch of aom, but this has not yet been published to a release tag in that repo, yet.